feat: add buffer for plotting signal

+++ src/device/inrat_v1/decoder.py
@@ -15,7 +15,8 @@ def decode_ecg(raw_data: bytearray) -> (int, np.ndarray):
     offset += 4
 
     # decode ecg
-    ecg = np.zeros(Pkt.SamplesCountEcg, dtype=np.int32)
+    # ecg = np.zeros(Pkt.SamplesCountEcg, dtype=np.int32)
+    ecg = np.zeros(Pkt.SamplesCountEcg, dtype=np.float32)
     prev = 0
     for i in range(Pkt.SamplesCountEcg):
         if (code >> i) & 0x1 == 0x0:
